Added ROSByteConvertedQueue for queue size converted in bytes.
Added a class ROSByteConvertedQueue similar to ROSConvertedQueue where the conversion is implicitely made by the function based on the size in byte of the given message.  This class uses the topic_tools::ShapeShifter to take a generic ROS message as an input. Therefore, there's no need to know the message type at the compilation time. The classes is then restricted to use the size() method of the shapeshifter to compute the size of the message in bytes.

- Added ROSByteConvertedQueue for ConvertedQueue with generic message where the conversion is made in bytes.
- Added the dependency to topic_tools
- Added tests to validate the constructors and bad initialization
- Test of the conversion was made externally since ROS test has some difficulties to execute the subscriber callback mid test.