Refactored the metric computation service for a more float request.
Since the ROS queues might receive the size of another queue as a metric, a standard interface to compute the size of the queue and the metrics is thus defined. Therefore, FloatRquest returns a float that could be a metric or the size of a queue. Integer sized queues need to be aware that their size will be casted as a float.