Added asynchronous service calls to the queue_controller
Since there is no requirements to get the parameters for the control setep except for the action set, all the service calls could be done in parallel. To prevent to add even more message definitions by using actionlib, another approach is used that puts the services in threads and just waits for the results. The approach uses std::future.

- Added a structure AsyncMetricFutureStruct to interact with the asynchronous calls of the services.
- Added an enumeration to give a mechanism to know which type of parameter is being fetched.
- Added methods to create an asynchronous service call and add it to a future vector to wait for the results. One method is for services that use an action in their request and the other one for ros_queue_msgs::FloatRequest messages.
- Adapted the getParametersForControlStep() method so the penalty, renewal time, arrival and departures are called in parallel ant to wait for their results at the end.
-