Added the dynamic change of the metrics
The main idea behind the dynamic of the states that was implemented is that at each change of the commanded zone, the state managers takes a snapshot of the current state of the system and it restarts an internal timer. Each time, an interface wants to know the current state of the system, the last transition's states is used and augmented by the rate of change of the temperature times the elapsed time. Other metrics like the completion and the penalty is also scale based on this time. In other words, the stored states are only changed at transitions and the get method only extrapolate current states.

- Changed the name of state message in the state manager to the states_at_last_transition to be more relevant.
- Added a commandToNextZone method in the state manager as the interface to send a new command.
- Maded the internal pointers to the services in the AUVSystem public so it could be accessible by the state manager
- Added public methods to each metrics so the state manager could access it through the AUV system