Changed the model to remove displacements between zones.
Since it caused a problem for the experiments to "stop between zones" if the time the renewal time is shorter than expected, it's hard to manage how where the robot is. If it's in the target zone, than the robot could magically go faster if the controller is faster. So now we change the penalty for joule transition as a transition cost and we now assume that the robot can move very quickly between those states. As if it needed just to move around itself to change zones. The renewal time also becomes a time of exploiting the target zone and not the time of transitions.

- Changed the penalty to not be scaled by the remaining distance since its a transition cost.
- Changed the penalty in the AUVStates so it now contains the total energy spent for transitions