Removed jitter and delay noise in monitoring by moving the publish auv state before applying action
It was noticed that the difference real_continuous_average_diff_with_server_time_average for the localization had some zig-zags in it for the step inverted controllers. It came mainly from the delay between monitoring that measure the system at the end of the control step but the queue server was only updated at the start of the next controller step.

However, the temperature was also affected in non inversion case since the queue server was updated but the auv system didn't have enough time so let its temperature rise to the change amount. But the error was sneaky since it was smoothed by the dynamical system.

Solution: Put the publishing of the auv state that triggers the monitoring each time the auv system received a command but before applying that command