Moved the monitoring sample between the controller start even and optimization done event
There was a problem where the changes estiamted by the queue server where not syncrhonised with when the monitoring node took its capture of the states. The difference between the real states and the estimation of the controller were thus lagged by those unsyncrhonized event.

To get the right temperature, the monitoring needs to wait for a whole control step to know what truly happened (state propagataion). On the other hand, he non-inversion of the controller requires the monitoring to be done before it ends its control loop to capture the what happend during the last frame. Thos constraints required a change on when to sample the monitoring.

The multicontroller syncrhonization signals of the optimal controller were used for those timings. The monitoring sample is taken at the beginning of the control step based on the signal "control_loop_started" and the end of the sampling sends a signal to indicate that the optimal controller can send its command. Since the queue_controller is dependent on the optimal_controller and the optimal_controller is dependent on the sampling signal, there is cascading dependency where the we ensure that the queue_controller doesn't send it's update to the queue_sever before the sampling is done.

- Added the multicontroller dependency of the optimal controller to the monitoring task in its config file.
- Created a new class VirtualQueueController that holld the metric_monitor object of the localization and temperaature. This new class susbcribes to the auv_stats topics and called the callbacks of the localization and temperature callbacks. It also provides a the queue_server stats to those callbacks with a service client call to the queue server.
- Removed the subscribers from the localization and temperature monitors