Added the reading of the queue sizes from the queue server from the distruber action server
Initialy the synchronized topic from the disturbed action server and from the optimal controller's action were used for the logging of the graphs. However, it was discovered that the synchronization through the message filter package was delayed by a whole cycle compared to the queue server. It was thus not usable for a graph of the queue size and the taken action. Instead of trying to remove this delay, this commit add a patch by addind the reading of the queue server when the action of the disturbed is published.