Wait for the existence of the services for the queues at the first update to prevent system deadlock
Since the queue controller is dependent on the existence of services of the queue server, the system might be dependent on the existence of the queue controller interfaces and that the queue server might be dependent on services from the system, waiting for the existence of another service might cause a system deadlock at initialization. This commit has the purpose to fix that by not causing the queue server to be dependent on other services at its initialization to create all its services. The dependence is only checked a the first update.
- Moved the wait for existence at the first update call for all ros_queues.
- Added the persistent service to the conversion_service_client of the ROSConvertedQueue class.